Moved download model to model package. Removed obsolete experimental download driver.
