+++ src/main/java/de/bayern/gdi/model/DownloadStep.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package de.bayern.gdi.utils;
+package de.bayern.gdi.model;
 
 import java.io.BufferedInputStream;
 import java.io.BufferedOutputStream;
